Update for LongColumn need to do some perf testing to see the impact of long vs int.

+++ src/main/java/ca/junctionbox/bullet/Applicable.java
@@ -10,5 +10,5 @@ public interface Applicable {
      *
      * @param i - the integer value.
      */
-    void each(final int i);
+    void each(final long i);
 }
